backend/requirements.txt
Flask
flask-cors
openai-whisper 
# or whisperx if you want alignment/timestamps
